oo-accept-broker: RHEL6 compatibility
oo-accept-broker is written for Fedora, which uses systemd.  In
particular, check_auth_mongo is written with the assumption that the
service command will print systemd-style output.  These assumptions
break on RHEL6 where the service command prints sysvinit-style output.

This commit adds compatibility so that oo-accept-broker should parse the
output of the service command correctly on both recent Fedora and RHEL6.

---------
This was merged into origin-server in commit
1e1350f16728bd4efd79018ef546def9b8c19adc