BZ886159 - Moving the console port from 3128 to 8118
This is a change that we are making to stay in sync with upstream.  The Fedora
selinux policy does not allow apache to bind to port 3128.  Both Fedora and
RHEL allow 8118.  You can verify this by running the following command:

semanage port -l | grep http

NOTE: I made a number of other changes to the console's spec file to properly
label the configuration files as noreplace.