Bug 902630: fix `service openshift-broker reload`
In the initscript, the options to give to httpd are stored in the
$OPTIONS variable.  In the current version of the script, the status
action has an error that causes $OPTIONS to be expanded incorrectly.
This commit resolves the issue by using eval, set, and "$@" to coax BASH
into performing the expansion we want.

This commit also fixes an unrelated issue: The reload action did not
print anything on success, and now it will.

This commit should fix Bug 902630.